Move mtgFilter into separate folder
