+++ src/reducers/colorReducer.js
@@ -1,16 +1,16 @@
-import { GET_ALL_CARDS } from '../actions/types';
-const initialState = '';
+import { GET_ALL_CARDS } from '../actions/filter/types'
+const initialState = ''
 
 const colorReducer = (state = initialState, action) => {
-  Object.freeze(state);
-  let newState;
+  Object.freeze(state)
+  let newState
   switch (action.type) {
     case GET_ALL_CARDS:
-      newState = action.colors;
-      return newState;
+      newState = action.colors
+      return newState
     default:
-      return state;
+      return state
   }
-};
+}
 
-export default colorReducer;
+export default colorReducer
